Concat some pyxl original test files
