+++ tests/pyxl/original/test_attr_name_case.py
@@ -1,4 +1,6 @@
 # coding: mixt
 from mixt import html
+
+
 def test():
     assert str(<div cLaSs="foo"></div>) == '<div class="foo"></div>'
