--- tests/pyxl/original/test_curlies_in_attrs_2.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment><img src="barbaz{'foo'}" /></Fragment>) == """<img src="barbazfoo" />"""
