--- tests/pyxl/original/test_curlies_in_strings_1.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment> '{'foobar'}' </Fragment>) == """ 'foobar' """
