--- tests/pyxl/original/test_curlies_in_strings_3.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment> "{' "foobar" '}" </Fragment>) == ''' " &quot;foobar&quot; " '''
