--- tests/pyxl/original/test_curlies_in_strings_4.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>"</Fragment>) + '{}' == '''"{}'''
