+++ tests/pyxl/original/test_eof.py
@@ -1,4 +1,6 @@
 # coding: mixt
 from mixt import html
+
+
 def test():
     assert str(<Fragment>'''</Fragment>) == """'''"""
