+++ tests/pyxl/original/test_html_comments.py
@@ -1,4 +1,10 @@
 # coding: mixt
 from mixt import html
-def test():
+
+
+def test1():
     assert str(<Fragment><!-- comment here --></Fragment>) == ""
+
+
+def test2():
+    assert str(<Fragment><!-- comment-here --></Fragment>) == ""
