--- tests/pyxl/original/test_html_comments_2.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment><!-- comment-here --></Fragment>) == ""
