+++ tests/pyxl/original/test_if.py
@@ -2,6 +2,22 @@
 from mixt import html
 
 def test():
+    assert str(<Fragment><if cond="{True}">true</if><else>false</else></Fragment>) == "true"
+    assert str(<Fragment><if cond="{False}">true</if><else>false</else></Fragment>) == "false"
+
+
+def test2():
+    assert str(<Fragment>
+                   <if cond="{True}">true</if>
+                   <else>false</else>
+               </Fragment>) == "true"
+    assert str(<Fragment>
+                   <if cond="{False}">true</if>
+                   <else>false</else>
+               </Fragment>) == "false"
+
+
+def test3():
     assert str(<Fragment>
                    <if cond="{True}">
                        <if cond="{True}">
@@ -77,3 +93,22 @@ def test():
                        </else>
                    </else>
                </Fragment>) == "four"
+
+
+def test4():
+    count = [0]
+    def foo(value):
+        count[0] += 1
+        return value
+    assert str(<Fragment>
+                   <if cond="{foo(True)}">a</if>
+                   <else>b</else>
+                   {count[0]}
+               </Fragment>) == "a1"
+
+    count[0] = 0
+    assert str(<Fragment>
+                   <if cond="{foo(False)}">a</if>
+                   <else>b</else>
+                   {count[0]}
+               </Fragment>) == "b1"
