--- tests/pyxl/original/test_if_2.py
@@ -1,12 +1,0 @@
-# coding: mixt
-from mixt import html
-
-def test():
-    assert str(<Fragment>
-                   <if cond="{True}">true</if>
-                   <else>false</else>
-               </Fragment>) == "true"
-    assert str(<Fragment>
-                   <if cond="{False}">true</if>
-                   <else>false</else>
-               </Fragment>) == "false"
