--- tests/pyxl/original/test_python_comments_1.py
@@ -1,5 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>Im cool # lol
-</Fragment>) == """Im cool """
