--- tests/pyxl/original/test_python_comments_3.py
@@ -1,5 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div #style="display: none;"
-               ></div>) == "<div></div>"
