--- tests/pyxl/original/test_tags_in_curlies_1.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{'<br />'}</Fragment>) == """&lt;br /&gt;"""
