--- tests/pyxl/original/test_tags_in_curlies_2.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{'<img src="foo" />'}</Fragment>) == """&lt;img src=&quot;foo&quot; /&gt;"""
