--- tests/pyxl/original/test_tags_in_curlies_3.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{'<div> foobar </div>'}</Fragment>) == """&lt;div&gt; foobar &lt;/div&gt;"""
