--- tests/pyxl/original/test_tags_in_curlies_4.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{'<div class="foo"> foobar </div>'}</Fragment>) == """&lt;div class=&quot;foo&quot;&gt; foobar &lt;/div&gt;"""
