--- tests/pyxl/original/test_tags_in_curlies_5.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment> {'<img src="{cond}" />'} </Fragment>) == """ &lt;img src=&quot;{cond}&quot; /&gt; """
