--- tests/pyxl/original/test_tags_in_curlies_6.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment> {' "<br /> '} </Fragment>) == '''  &quot;&lt;br /&gt;  '''
