--- tests/pyxl/original/test_tags_in_curlies_9.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{<br /> if True else <div></div>}</Fragment>) == '''<br />'''
