--- tests/pyxl/original/test_whitespace_1.py
@@ -1,6 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div class="{'blah'}">
-                   blah <a href="%(url)s">blah</a> blah.
-               </div>) == """<div class="blah">blah <a href="%(url)s">blah</a> blah.</div>"""
