--- tests/pyxl/original/test_whitespace_10.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div class="{'foo'} {'bar'}"></div>) == '<div class="foo bar"></div>'
