--- tests/pyxl/original/test_whitespace_2.py
@@ -1,6 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div>
-                   The owner has not granted you access to this file.
-               </div>) == """<div>The owner has not granted you access to this file.</div>"""
