--- tests/pyxl/original/test_whitespace_3.py
@@ -1,9 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    a = (<br />)
-    b = (<div>
-             foo
-         </div>)
-    assert str(b) == "<div>foo</div>"
-    assert a  # pacify lint
