--- tests/pyxl/original/test_whitespace_4.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div class="{ 'foo' }">foo</div>) == '<div class="foo">foo</div>'
