--- tests/pyxl/original/test_whitespace_5.py
@@ -1,7 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>
-                   {'foo'}
-                   {'bar'}
-               </Fragment>) == "foo bar"
