--- tests/pyxl/original/test_whitespace_6.py
@@ -1,9 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>
-                   {'foo'}
-                   <if cond="{True}">
-                       {'foo'}
-                   </if>
-               </Fragment>) == "foofoo"
