--- tests/pyxl/original/test_whitespace_7.py
@@ -1,7 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>
-                   foo
-                   {'foo'}
-               </Fragment>) == "foo foo"
