--- tests/pyxl/original/test_whitespace_8.py
@@ -1,4 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<Fragment>{ 'foo' }{ 'foo' }</Fragment>) == "foofoo"
