--- tests/pyxl/original/test_whitespace_9.py
@@ -1,6 +1,0 @@
-# coding: mixt
-from mixt import html
-def test():
-    assert str(<div class="foo
-                           bar">
-               </div>) == '<div class="foo bar"></div>'
